Interceptor로 권한 확인 후 기능 수행 및 파라미터 유효성 체크 수행

+++ src/main/java/com/ara/todayoutfit/recommend/model/RecommendInfo.java
@@ -15,7 +15,7 @@ public class RecommendInfo implements Serializable {
     @Id
     @Column(name="recommendinfo_seq")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer recommendInfoSeq;
+    private Long recommendInfoSeq;
 
     @Column(name="max_temp")
     private Integer maxTemp;
